requirements.txt
streamlit>=1.34
polars>=0.20
numpyro>=0.13
scikit-learn>=1.4
dask[distributed]>=2024.4
celery>=5.3
redis>=5.0
altair>=5.2
weasyprint
pytest
pytest-asyncio
hypothesis

# Install your own code as a package
-e .
